Update adventureEngine.py
The example crashes when trying to run and issuing commands such as "look" because of changes in pyparser. Return parseCmd is now returning ParseResult and not an object with a command method that can be called with arg p. This is a quick fix to make the example work, it is not clear from the documentation why return now differs.